bugfixing

+++ C/sleep.c
@@ -2,10 +2,11 @@
 
 #define addr 0x20003004
 
+//Not working
 void Sleep(int mcsec)
 {
-    uint8_t timer_start = *(int *)addr;
-    uint8_t timer_current = *(int *)addr;
+    uint32_t timer_start = *(int *)addr;
+    uint32_t timer_current = *(int *)addr;
     while(timer_start + mcsec <= timer_current)
     {
         timer_current = *(int *)addr;
